+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 ## Unreleased
 
+- **Breaking**:
+    - Requires Clojure 1.9 (for spec)
+- Sass4clj now contains main namespace for `clj` use
+- Sass4clj new has `sass4clj.api` namespace with easy to use `start` and `stop` functions
+- Add [Integrant](https://github.com/weavejester/integrant) namespace `sass4clj.integrant` namespace
+- Add [Component](https://github.com/stuartsierra/component) namespace `sass4clj.component` namespace
+- Use [Hawk](https://github.com/wkf/hawk/) for watching for file changes, this should work better on OS X
 - Support Webpack style import paths, which support starting the path with `~`
 when referring to Node packages.